chore: code review feedback
