+++ Cargo.toml
@@ -30,7 +30,6 @@ automations = { path = "./crates/automations" }
 async-trait = "0.1.51"
 serde = { version = "1.0.130", features = ["derive"] }
 logger = { path = "./crates/logger" }
-webhooks = { path = "./crates/webhooks" }
 serde_json = "1.0.68"
 colored = "2.0.0"
 base64 = "0.13.0"
