+++ crates/cli/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 afj-rest = { path = "../afj-rest" }
 cloudagent-python = { path = "../cloudagent-python" }
 agent = { path = "../agent" }
-webhooks = { path = "../webhooks" }
 automations = { path = "../automations" }
 logger = { path = "../logger" }
 base64 = "0.13.0"
@@ -19,3 +18,5 @@ serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
 serde_yaml = "0.8.23"
 tokio = { version = "1", features = ["full"] }
+tungstenite = {version = "0.17.3", features = ["native-tls"]}
+url = "2.2.2"
