+++ crates/cli/src/register.rs
@@ -5,13 +5,13 @@ use crate::modules::basic_message::parse_basic_message_args;
 use crate::modules::configuration::parse_configuration_args;
 use crate::modules::connection::parse_connection_args;
 use crate::modules::oob::parse_oob_args;
-use crate::modules::webhooks::parse_webhooks_args;
 use crate::modules::credential::parse_credentials_args;
 use crate::modules::credential_definition::parse_credential_definition_args;
 use crate::modules::feature::parse_features_args;
 use crate::modules::multitenancy::parse_multitenancy_args;
 use crate::modules::proof::parse_proof_args;
 use crate::modules::schema::parse_schema_args;
+use crate::modules::webhooks::listen;
 use crate::utils::config::{get_config_from_path, get_config_path};
 use afj_rest::agent::{CloudAgentAfjRest, CloudAgentAfjRestVersion};
 use clap::Parser;
@@ -65,7 +65,7 @@ pub async fn register() -> Result<()> {
                 match &cli.commands {
                     Commands::Schema(options) => parse_schema_args(options, agent).await,
                     Commands::Feature(_) => parse_features_args(agent).await,
-                    Commands::Webhooks(options) => parse_webhooks_args(options).await,
+                    Commands::Webhooks(_) => {listen(agent.endpoint)},
                     Commands::Message(options) => parse_basic_message_args(options, agent).await,
                     Commands::CredentialDefinition(options) => {
                         parse_credential_definition_args(options, agent).await
