--- crates/webhooks/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "webhooks"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tungstenite = {version = "0.17.3", features = ["native-tls"]}
-url = "2.2.2"
-serde_json = "1.0.68"
-colored = "2.0.0"
